refactor: #189 - Spinner 경로 수정
