+++ src/components/SpaceList/SpaceList.tsx
@@ -4,8 +4,8 @@ import { Fragment } from 'react'
 import { CATEGORIES_RENDER } from '@/constants'
 import useInfiniteScroll from '@/hooks/useInfiniteScroll'
 import { SearchSpaceReqBody, SpaceResBody } from '@/types'
+import { Spinner } from '..'
 import Space from '../common/Space/Space'
-import Spinner from '../common/Spinner/Spinner'
 import { NONE_SEARCH_RESULT } from './constants'
 import useSpacesQuery from './hooks/useSpacesQuery'
