+++ src/components/common/Sidebar/Sidebar.tsx
@@ -1,13 +1,13 @@
 'use client'
 
 import { useRef } from 'react'
+import { Spinner } from '@/components'
 import { useCurrentUser } from '@/hooks/useCurrentUser'
 import { XMarkIcon } from '@heroicons/react/24/outline'
 import { ArchiveBoxIcon, StarIcon } from '@heroicons/react/24/solid'
 import Link from 'next/link'
 import Avatar from '../Avatar/Avatar'
 import Button from '../Button/Button'
-import Spinner from '../Spinner/Spinner'
 import ThemeButton from '../ThemeButton/ThemeButton'
 import { useMySpace } from './hooks/useMySpace'
 import useSidebar from './hooks/useSidebar'
